pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qrank"
version = "0.1.0"
description = "Analyze and rank SQL queries based on efficiency and best practices"
authors = [{name = "Carlos", email = "seuemail@exemplo.com"}]
dependencies = ["sqlparse", "psycopg2-binary", "mysql-connector-python", "click"]
readme = "README.md"
requires-python = ">=3.8"

[project.scripts]
qrank = "qrank.__main__:main"